+++ packages/core-data/src/entity-types/helpers.ts
@@ -18,7 +18,15 @@ export interface AvatarUrls {
 export type MediaType = 'image' | 'file';
 export type CommentingStatus = 'open' | 'closed';
 export type PingStatus = 'open' | 'closed';
-export type PostStatus = 'publish' | 'future' | 'draft' | 'pending' | 'private';
+export type PostStatus =
+	| 'publish'
+	| 'future'
+	| 'draft'
+	| 'pending'
+	| 'private'
+	| 'auto-draft'
+	| 'trash'
+	| ( string & {} );
 export type PostFormat =
 	| 'standard'
 	| 'aside'