+++ packages/core-data/src/entity-types/type.ts
@@ -39,7 +39,11 @@ declare module './base-entity-records' {
 			/**
 			 * All features, supported by the post type.
 			 */
-			supports: ContextualField< Record< string, string >, 'edit', C >;
+			supports: ContextualField<
+				Record< string, true | unknown[] >,
+				'edit',
+				C
+			>;
 			/**
 			 * Taxonomies associated with post type.
 			 */