+++ packages/core-data/src/entity-types/wp-template-part.ts
@@ -77,7 +77,7 @@ declare module './base-entity-records' {
 			/**
 			 * Theme file exists.
 			 */
-			has_theme_file: Record< string, string >;
+			has_theme_file: boolean;
 			/**
 			 * The ID for the author of the template.
 			 */